Change method for adding cards to database etc.
Changed the method for adding a card created by another user to ones own
collection. Changed userId in the card model to an array of userIds, so
that rather than replicating the card, the users id is added to the
existing card. Also updated the design and UI, for aesthetics and ease-
of-use. (if not mentioned in previous push, also added a card rating
system)