Add AI audio generation, implement revision feature
Added an AI audio generation option for the kunyomi and onyomi in card
creation. Began implementing the revision feature, by adding interval and
lastStudied components to model, if more time than the set interval has
passed since the card was last studied, it will be presented for
revision. Will add a patch request for updating these values, depending
on whether the card was answered correctly or not.