Convert to a web app structure seperate to landing page
Fundamentally shifted the design of the application from a website structure to a web application structure. Made the profile page the main page and added navigation buttons to the dashboard. Made the navbar slimmer to prevent distraction. Also structured pages to prevent uneccessary scrolling for a focussed environment. Improved general ui/ux.